+++ src/commands/check.ts
@@ -1,6 +1,7 @@
 import { Project } from '../common/Config';
 import { PipelineCommand } from '../common/pipelines';
-import { StatusCode } from '../common/api';
+import { Response, StatusCode } from '../common/api';
+import { IPipeline } from '../common/iPipeline';
 
 export class Check extends PipelineCommand {
   protected async runPerProject(project: Project) {
@@ -8,19 +9,16 @@ export class Check extends PipelineCommand {
     if (this.yargs.await) {
       return await this.awaitPipelineCompletion(project, pipeline.data);
     } else {
-      return await this.checkStatus(project);
+      return await this.checkStatus(project, pipeline);
     }
   }
 
-  private async checkStatus(project: Project) {
-    this.screenPrinter.setProjectSpinner(project, 'Searching pipeline...');
-
-    const resp = await this.getPipeline(project, this.yargs.ref);
-    const data = resp.data;
-    if (resp.status === StatusCode.Success) {
+  private async checkStatus(project: Project, pipeline: Response<IPipeline>) {
+    const data = pipeline.data;
+    if (pipeline.status === StatusCode.Success) {
       const msg = `Pipeline status: ${data.status} last update ${data.created_at}`;
       this.screenPrinter.setProjectSuccess(project, msg);
     }
-    return resp;
+    return pipeline;
   }
 }